+++ satella/instrumentation/metrics/metric_types/__init__.py
@@ -1,12 +1,11 @@
 from .base import Metric, LeafMetric
 from .cps import ClicksPerTimeUnitMetric
-from .simple import StringMetric, IntegerMetric, FloatMetric
+from .simple import IntegerMetric, FloatMetric
 from .percentile import PercentileMetric
 
 ALL_METRICS = [
     Metric,
     LeafMetric,
-    StringMetric,
     IntegerMetric,
     FloatMetric,
     ClicksPerTimeUnitMetric,